refactor(web): split session UI components
Extract the live session card into protocol-specific section components and move shared formatting into a dedicated helper module.

Reuse shared chart panel and zoom modal wrappers inside SessionCharts so the repeated chart shell and modal controls are no longer duplicated inline.

Update ABSOLUTE_REVIEW_TODO.md so the large-scope storage, sampler, and session UI refactors are recorded with their current completion status.
